Replaced and removed old form renderer.
