+++ src/SmartComponents/Forms/DataDrivenForm/index.js
@@ -1,1 +1,2 @@
+export * from '../FormFields/formFields';
 export { default as FormRenderer } from './formRenderer';
