--- src/SmartComponents/Forms/DataDrivenForm/validationMapper.js
@@ -1,6 +1,0 @@
-import { required, minLength } from '../Validators/validators';
-
-export default validationType => ({
-    required,
-    minLength
-})[validationType];
