--- src/SmartComponents/Forms/DataDrivenForm2/index.js
@@ -1,1 +1,0 @@
-export * from './formFields';
