+++ src/SmartComponents/Forms/index.js
@@ -1,1 +1,1 @@
-export { FormRenderer } from './DataDrivenForm';
+export * from './DataDrivenForm';
